Migrated build to root directory. * Now emit nice warning if scala-dist.zip is not found. * Can share scala-dist.zip accross projects * Can share builds when making distribution.
